requirements.txt
django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
mysqlclient==2.2.4
django-redis==5.4.0
pycryptodomex==3.20.0
djangorestframework-simplejwt==5.3.1
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
flower==2.0.1
python-daemon==3.0.1
gunicorn==22.0.0
django-proxy==1.3.0
psutil==5.9.8
uvicorn==0.29.0
daphne==4.1.2
channels==4.1.0
channels-redis==4.2.0
django-simple-captcha==0.6.0
user-agents==2.2.0
aiofiles==23.2.1
websockets==12.0
django-imagekit==5.0.0
pilkit==3.0
drf-yasg==1.21.7
requests==2.31.0